requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
huggingface-hub==0.16.4
pdfplumber==0.10.3
python-docx==0.8.11
langdetect==1.0.9
wikipedia-api==0.6.0
requests==2.31.0
zipfile36==0.1.3
transformers==4.34.0
torch==2.0.1